Moved Signup.tsx  Added authentication service code
