+++ src/ThePlatformWebSite.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import HomePage from './pages/HomePage';
 import Organization from './pages/OrganizationPage';
+import SignupPage from './pages/SignupPage';
 
 import './App.scss';
 
@@ -13,6 +14,7 @@ export default function ThePlatformWebsite( props:{} ){
             <BrowserRouter>
                 <Switch>
                     <Route path="/organization" component={Organization} />
+                    <Route path="/signup" component={SignupPage} />
                     <Route path="/" component={HomePage} />
                 </Switch>
             </BrowserRouter>
