+++ src/pages/style.scss
@@ -1,6 +1,6 @@
 .signup-ctr{
     border: 3px;
-    width: 30%;
+    width: 70%;
     margin: auto;
     height: auto;
 }
@@ -22,4 +22,7 @@
 .submit{
     padding-top: 5px;
     margin-top: 5px;
+    width: 20%;
+    margin-left: auto;
+    display: block;
 }
